+++ modules/download-link/download-link.hpp
@@ -25,7 +25,7 @@
  *
  *******************************************************************************/
 
- /* $Id: download-link.hpp 1686 2005-07-24 01:10:54Z common $ */
+ /* $Id: download-link.hpp 1926 2005-08-27 20:52:47Z dp $ */
 
 #ifndef HAVE_CSENDDOWNLOADHANDLER_HPP
 #define HAVE_CSENDDOWNLOADHANDLER_HPP
@@ -46,8 +46,8 @@ namespace nepenthes
 {
 	struct LinkBindContext
 	{
-		unsigned long 		m_RemoteHost;
-		unsigned short int 	m_LocalPort;
+		uint32_t 		m_RemoteHost;
+		uint16_t 	m_LocalPort;
 		string				m_Challenge;
 		Download 			*m_Download;
 	};
@@ -69,8 +69,8 @@ namespace nepenthes
 		void socketClosed(Socket *socket);
 
 	protected:
-		unsigned int m_MaxFileSize;
-		unsigned int m_ConnectTimeout;
+		uint32_t m_MaxFileSize;
+		uint32_t m_ConnectTimeout;
 
 		list <LinkBindContext *> m_BindContexts;
